Add utlis file to hold common functions for DRYer code
