+++ src/Components/MovieMainPage/MovieMainPage.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
-import MovieRatingForm from "../MovieRatingForm/MovieRatingForm"
 import "./MovieMainPage.css";
 import { getSingleMovieInfo } from "../../apiCalls";
+import { formatDate, findMovieRating } from "../../utils";
 
 
 class MovieMainPage extends Component {
@@ -15,12 +15,6 @@ class MovieMainPage extends Component {
       }
     };
     return getMovieDetails();
-  }
-
-  findMovieRating (movie)  {
-    if (!this.props.ratings.length) return <MovieRatingForm userId={this.props.ratings[0].user_id} movieId={movie.id}/>
-    const movieRating = this.props.ratings.find( rating => movie.id === rating.movie_id)
-    return (movieRating) ? <MovieRatingForm userId={this.props.ratings[0].user_id} movieId={movie.id} userRating={movieRating}/> : <MovieRatingForm userId={this.props.ratings[0].user_id} movieId={movie.id} />
   }
 
   render() {    
@@ -38,13 +32,13 @@ class MovieMainPage extends Component {
                 <h3>{this.state.movie.title}</h3>
                 <h4 className="tagline">{this.state.movie.tagline}</h4>
                 <p><strong>Overview:</strong> <br></br> {this.state.movie.overview}</p>
-                <p><strong>Release Data:</strong> <br></br> {this.state.movie.release_date}</p>
+                <p><strong>Release Data:</strong> <br></br> {formatDate(this.state.movie.release_date)}</p>
                 <p><strong>Genre(s):</strong> <br></br> {(this.state.movie.genres).join(', ')}</p>
                 <p><strong>Budget</strong>: <br></br> ${(this.state.movie.budget).toLocaleString('en')}</p>
                 <p><strong>Revenue:</strong> <br></br> ${(this.state.movie.revenue).toLocaleString('en')}</p>
                 <p><strong>Runtime:</strong> <br></br> {this.state.movie.runtime} Minutes</p>
                 <p><strong>Average Rating:</strong> <br></br> {Math.round(this.state.movie.average_rating)}</p>
-                {(this.props.ratings) ? this.findMovieRating(this.state.movie) : null}
+                {(this.props.ratings) ? findMovieRating(this.props.ratings, this.state.movie,this.props.ratings[0].user_id) : null}
               </section>
             </figure>
           </section>
