+++ src/Components/Movies/Movies.js
@@ -2,15 +2,9 @@ import React from "react";
 import Movie from "../Movie/Movie.js";
 import "./Movies.css";
 import { Link } from "react-router-dom";
-import MovieRatingForm from "../MovieRatingForm/MovieRatingForm"
+import { findMovieRating } from "../../utils";
 
 const Movies = ({ movies, ratings, userId }) => {
-  const findMovieRating = (movie) => {
-    if (!ratings.length) return <MovieRatingForm userId={userId} movieId={movie.id}/>
-    const movieRating = ratings.find( rating => movie.id === rating.movie_id)
-    return (movieRating) ? <MovieRatingForm userId={userId} movieId={movie.id} userRating={movieRating}/> : <MovieRatingForm userId={userId} movieId={movie.id} />
-  }
-
   const movieResults = movies.map((movie, index) => {
     return (
       <li id={movie.id} key={index}>
@@ -23,7 +17,7 @@ const Movies = ({ movies, ratings, userId }) => {
             releaseDate={movie.release_date}
           />
         </Link>
-        {(ratings) ? findMovieRating(movie) : null}
+        {(ratings) ? findMovieRating(ratings, movie, userId) : null}
       </li>
     );
   });
